requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
prometheus-fastapi-instrumentator
pyfunctional==1.4.3
pyarango==1.3.5